backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pytest==7.4.0
gunicorn==21.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
